crm/requirements.txt
google-generativeai>=0.4.1
playwright==1.42.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
requests>=2.31.0
dnspython>=2.4.0
psutil>=5.9.0
apscheduler>=3.10.0
pytz>=2024.1
lxml>=5.0.0
